Move IComparable assertions to extension methods
