+++ src/Libraries/AssertNet/AssertionTypes/Objects/DoubleAssertion.cs
@@ -17,90 +17,6 @@ public class DoubleAssertion : Assertion<DoubleAssertion, double>
     }
 
     /// <summary>
-    /// Asserts if a double is greater than a certain value.
-    /// </summary>
-    /// <param name="other">Value which the double should be greater than.</param>
-    /// <param name="message">Custom message for the assertion failure.</param>
-    /// <returns>The current assertion.</returns>
-    [Assertion]
-    public DoubleAssertion IsGreaterThan(double other, string? message = null)
-    {
-        if (Subject <= other)
-        {
-            this.Fail(new FailureBuilder("IsGreaterThan()")
-                .Append(message)
-                .Append("Expecting", Subject)
-                .Append("To be greater than", other)
-                .Finish());
-        }
-
-        return this;
-    }
-
-    /// <summary>
-    /// Asserts if a double is greater than or equal to a certain value.
-    /// </summary>
-    /// <param name="other">Value which the double should be greater than or equal to.</param>
-    /// <param name="message">Custom message for the assertion failure.</param>
-    /// <returns>The current assertion.</returns>
-    [Assertion]
-    public DoubleAssertion IsGreaterThanOrEqualTo(double other, string? message = null)
-    {
-        if (Subject < other)
-        {
-            this.Fail(new FailureBuilder("IsGreaterThanOrEqualTo()")
-                .Append(message)
-                .Append("Expecting", Subject)
-                .Append("To be greater than or equal to", other)
-                .Finish());
-        }
-
-        return this;
-    }
-
-    /// <summary>
-    /// Asserts if a double is lesser than a certain value.
-    /// </summary>
-    /// <param name="other">Value which the double should be lesser than.</param>
-    /// <param name="message">Custom message for the assertion failure.</param>
-    /// <returns>The current assertion.</returns>
-    [Assertion]
-    public DoubleAssertion IsLesserThan(double other, string? message = null)
-    {
-        if (Subject >= other)
-        {
-            this.Fail(new FailureBuilder("IsLesserThan()")
-                .Append(message)
-                .Append("Expecting", Subject)
-                .Append("To be lesser than", other)
-                .Finish());
-        }
-
-        return this;
-    }
-
-    /// <summary>
-    /// Asserts if a double is lesser than or equal to a certain value.
-    /// </summary>
-    /// <param name="other">Value which the double should be lesser than or equal to.</param>
-    /// <param name="message">Custom message for the assertion failure.</param>
-    /// <returns>The current assertion.</returns>
-    [Assertion]
-    public DoubleAssertion IsLesserThanOrEqualTo(double other, string? message = null)
-    {
-        if (Subject > other)
-        {
-            this.Fail(new FailureBuilder("IsLesserThanOrEqualTo()")
-                .Append(message)
-                .Append("Expecting", Subject)
-                .Append("To be lesser than or equal to", other)
-                .Finish());
-        }
-
-        return this;
-    }
-
-    /// <summary>
     /// Asserts if a double is equal to zero.
     /// </summary>
     /// <param name="message">Custom message for the assertion failure.</param>
