+++ src/Tests/AssertNet.Tests/AssertionTypes/Objects/DoubleAssertionTests.cs
@@ -19,86 +19,6 @@ public class DoubleAssertionTests : ObjectAssertionTests<DoubleAssertion, double
     }
 
     /// <summary>
-    /// Checks that the assertion passes if the value is greater than a certain value.
-    /// </summary>
-    [Fact]
-    public void IsGreaterThanPassTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 20).IsGreaterThan(10);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Never());
-    }
-
-    /// <summary>
-    /// Checks that the assertion fails if the value is not greater than a certain value.
-    /// </summary>
-    [Fact]
-    public void IsGreaterThanFailTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 15).IsGreaterThan(15);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Once());
-    }
-
-    /// <summary>
-    /// Checks that the assertion passes if the value is greater than or equal to a certain value.
-    /// </summary>
-    [Fact]
-    public void IsGreaterThanOrEqualPassTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 32).IsGreaterThanOrEqualTo(32);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Never());
-    }
-
-    /// <summary>
-    /// Checks that the assertion fails if the value is not greater than or equal to a certain value.
-    /// </summary>
-    [Fact]
-    public void IsGreaterThanOrEqualFailTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 2).IsGreaterThanOrEqualTo(23);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Once());
-    }
-
-    /// <summary>
-    /// Checks that the assertion passes if the value is lesser than a certain value.
-    /// </summary>
-    [Fact]
-    public void IsLesserThanPassTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 12).IsLesserThan(13);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Never());
-    }
-
-    /// <summary>
-    /// Checks that the assertion fails if the value is not lesser than a certain value.
-    /// </summary>
-    [Fact]
-    public void IsLesserThanFailTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 25).IsLesserThan(10);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Once());
-    }
-
-    /// <summary>
-    /// Checks that the assertion passes if the value is lesser than or equal to a certain value.
-    /// </summary>
-    [Fact]
-    public void IsLesserThanOrEqualPassTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 2).IsLesserThanOrEqualTo(2);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Never());
-    }
-
-    /// <summary>
-    /// Checks that the assertion fails if the value is not lesser than or equal to a certain value.
-    /// </summary>
-    [Fact]
-    public void IsLesserThanOrEqualFailTest()
-    {
-        new DoubleAssertion(FailureHandler.Object, 2).IsLesserThanOrEqualTo(0);
-        FailureHandler.Verify(x => x.Fail(It.IsAny<string>()), Times.Once());
-    }
-
-    /// <summary>
     /// Checks that the assertion passes if the value is equal to 0.
     /// </summary>
     [Fact]
